requirements.txt
fastapi
uvicorn[standard]
google-genai
pandas
numpy
pillow
python-multipart
opencv-python
scipy
python-dotenv